pyproject.toml
[project]
name = "freeduckduckgo-assist"
version = "1.5.0"
description = "Free GPT-4o mini API using DuckDuckGo chat services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "fastapi[standard]>=0.115.7",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.34.0",
]
